+++ Tests/Unit/MSFT_SqlServerConfiguration.Tests.ps1
@@ -60,19 +60,10 @@ $invalidOption = @{
     RestartTimeout = 120
 }
 
-## compile the SMO stub
-#Add-Type -Path (Join-Path -Path $script:moduleRoot -ChildPath 'Tests\Unit\Stubs\SMO.cs')
-
 try
 {
     Describe "$($script:DSCResourceName)\Get-TargetResource" {
-
-        Mock -CommandName New-VerboseMessage -ModuleName $script:DSCResourceName
-
-        Mock -CommandName New-TerminatingError -MockWith { $ErrorType } -ModuleName $script:DSCResourceName
-
         Context 'The system is not in the desired state' {
-
             Mock -CommandName Connect-SQL -MockWith {
                 $mock = New-Object -TypeName PSObject -Property @{
                     Configuration = @{
@@ -85,13 +76,13 @@ try
                     }
                 }
 
-                ## add the Alter method
+                # Add the Alter method.
                 $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
 
                 return $mock
             } -ModuleName $script:DSCResourceName -Verifiable
 
-            ## Get the current state
+            # Get the current state.
             $result = Get-TargetResource @desiredState
 
             It 'Should return the same values as passed' {
@@ -109,7 +100,6 @@ try
         }
 
         Context 'The system is in the desired state' {
-
             Mock -CommandName Connect-SQL -MockWith {
                 $mock = New-Object -TypeName PSObject -Property @{
                     Configuration = @{
@@ -122,13 +112,13 @@ try
                     }
                 }
 
-                ## add the Alter method
+                # Add the Alter method.
                 $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
 
                 return $mock
             } -ModuleName $script:DSCResourceName -Verifiable
 
-            ## Get the current state
+            # Get the current state.
             $result = Get-TargetResource @desiredState
 
             It 'Should return the same values as passed' {
@@ -145,8 +135,7 @@ try
             }
         }
 
-        Context 'Invalid data is supplied' {
-
+        Context 'Invalid option name is supplied' {
             Mock -CommandName Connect-SQL -MockWith {
                 $mock = New-Object -TypeName PSObject -Property @{
                     Configuration = @{
@@ -159,24 +148,20 @@ try
                     }
                 }
 
-                ## add the Alter method
+                # Add the Alter method.
                 $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
 
                 return $mock
             } -ModuleName $script:DSCResourceName -Verifiable
 
-            It 'Should call New-TerminatingError mock when a bad option name is specified' {
-                { Get-TargetResource @invalidOption } | Should -Throw 'ConfigurationOptionNotFound'
-                Assert-MockCalled -ModuleName $script:DSCResourceName -CommandName New-TerminatingError -Scope It -Times 1
-                Assert-MockCalled -ModuleName $script:DSCResourceName -CommandName Connect-SQL -Scope Context -Times 1
+            It 'Should throw the correct error message' {
+                $errorMessage = ($script:localizedData.ConfigurationOptionNotFound -f $invalidOption.OptionName)
+                { Get-TargetResource @invalidOption } | Should -Throw $errorMessage
             }
         }
     }
 
     Describe "$($script:DSCResourceName)\Test-TargetResource" {
-
-        Mock -CommandName New-VerboseMessage -ModuleName $script:DSCResourceName
-
         Mock -CommandName Connect-SQL -MockWith {
             $mock = New-Object -TypeName PSObject -Property @{
                 Configuration = @{
@@ -189,7 +174,7 @@ try
                 }
             }
 
-            ## add the Alter method
+            # Add the Alter method.
             $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
 
             return $mock
@@ -205,10 +190,7 @@ try
     }
 
     Describe "$($script:DSCResourceName)\Set-TargetResource" {
-        Mock -CommandName New-VerboseMessage -ModuleName $script:DSCResourceName
-
         Mock -CommandName New-TerminatingError -ModuleName $script:DSCResourceName
-
         Mock -CommandName Connect-SQL -MockWith {
             $mock = New-Object -TypeName PSObject -Property @{
                 Configuration = @{
@@ -222,7 +204,7 @@ try
                 }
             }
 
-            ## add the Alter method
+            # Add the Alter method.
             $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
 
             return $mock
@@ -241,15 +223,14 @@ try
                 }
             }
 
-            ## add the Alter method
+            # Add the Alter method.
             $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
 
             return $mock
         } -ModuleName $script:DSCResourceName -Verifiable -ParameterFilter { $SQLServer -eq 'CLU02' }
 
         Mock -CommandName Restart-SqlService -ModuleName $script:DSCResourceName -Verifiable
-
-        Mock -CommandName New-WarningMessage -ModuleName $script:DSCResourceName -Verifiable
+        Mock -CommandName Write-Warning -ModuleName $script:DSCResourceName -Verifiable
 
         Context 'Change the system to the desired state' {
             It 'Should not restart SQL for a dynamic option' {
@@ -265,12 +246,37 @@ try
             It 'Should warn about restart when required, but not requested' {
                 Set-TargetResource @desiredState
 
-                Assert-MockCalled -ModuleName $script:DSCResourceName -CommandName New-WarningMessage -Scope It -Times 1 -Exactly
+                Assert-MockCalled -ModuleName $script:DSCResourceName -CommandName Write-Warning -Scope It -Times 1 -Exactly
                 Assert-MockCalled -ModuleName $script:DSCResourceName -CommandName Restart-SqlService -Scope It -Times 0 -Exactly
             }
 
             It 'Should call Connect-SQL to get option values' {
                 Assert-MockCalled -ModuleName $script:DSCResourceName -CommandName Connect-SQL -Scope Context -Times 3
+            }
+        }
+
+        Context 'Invalid option name is supplied' {
+            Mock -CommandName Connect-SQL -MockWith {
+                $mock = New-Object -TypeName PSObject -Property @{
+                    Configuration = @{
+                        Properties = @(
+                            @{
+                                DisplayName = 'user connections'
+                                ConfigValue = 0
+                            }
+                        )
+                    }
+                }
+
+                # Add the Alter method.
+                $mock.Configuration | Add-Member -MemberType ScriptMethod -Name Alter -Value {}
+
+                return $mock
+            } -ModuleName $script:DSCResourceName -Verifiable
+
+            It 'Should throw the correct error message' {
+                $errorMessage = ($script:localizedData.ConfigurationOptionNotFound -f $invalidOption.OptionName)
+                { Set-TargetResource @invalidOption } | Should -Throw $errorMessage
             }
         }
     }